chore: updated container to hosting container
