--- src/components/bottomSheetContainer/index.ts
@@ -1,2 +1,0 @@
-export { default } from './BottomSheetContainer';
-export type { BottomSheetContainerProps } from './types';
