+++ src/components/bottomSheetHostingContainer/index.ts
@@ -1,0 +1,2 @@
+export { BottomSheetHostingContainer as default } from './BottomSheetHostingContainer';
+export type { BottomSheetHostingContainerProps } from './types';
