+++ src/components/bottomSheetHostingContainer/styles.ts
@@ -1,8 +1,5 @@
 import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
-  root: {
-    flex: 1,
-  },
   container: StyleSheet.absoluteFillObject,
 });
